+++ application/models/Template_model.php
@@ -20,6 +20,7 @@ class Template_model extends MY_Model {
 		$CI->load->model('template_background_model');
 
 		$bg = $CI->template_background_model->getById($bg_id);
+		if ( ! $bg) return;
 		// Only pair with photos that have defaults configured
 		$photos = $this->db->where('is_active', 1)->where('has_defaults', 1)
 			->get('template_photos')->result();
@@ -59,7 +60,7 @@ class Template_model extends MY_Model {
 		$CI->load->model('template_photo_model');
 
 		$photo = $CI->template_photo_model->getById($photo_id);
-
+		if ( ! $photo) return;
 		// Only pair with backgrounds that have defaults configured
 		$bgs = $this->db->where('is_active', 1)->where('has_defaults', 1)
 			->get('template_backgrounds')->result();